requirements.txt
# TechPulse — Streamlit Cloud
# (PyQt6 y TikTokApi eliminados — no compatibles con Linux cloud)

streamlit>=1.35.0

# Reddit
praw==7.7.1

# YouTube
google-api-python-client==2.136.0

# Web scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2

# Playwright — Amazon scraper (browsers se instalan en runtime vía packages.txt)
playwright==1.45.0

# Database
SQLAlchemy==2.0.31

# Scheduler
APScheduler==3.10.4

# AI
anthropic==0.34.2

# Config
python-dotenv==1.0.1

# Utils
python-dateutil==2.9.0
pandas>=2.0.0

# Google Trends
pytrends==4.9.2

# Google Ads API — Keyword Planner (opcional)
google-ads>=22.1.0
